refactor(cockpit): extract contract detection module
